+++ buildSrc/src/main/kotlin/Configurations.kt
@@ -37,67 +37,67 @@ object Versions {
     }
 
     object Sdk {
-        const val classpath = "4.1.2"
-        const val buildTools = "30.0.3"
+        const val classpath = "7.0.0"
+        const val buildTools = "31.0.0"
         const val compile = 30
         const val min = 21
         const val target = 30
     }
 
-    val java = JavaVersion.VERSION_1_8
+    val java = JavaVersion.VERSION_11
 
     object Kotlin {
-        const val jvmTarget = "1.8"
-        const val core = "1.4.30"
-        const val coroutines = "1.4.2"
+        const val jvmTarget = "11"
+        const val core = "1.5.21"
+        const val coroutines = "1.5.1"
     }
 
     object AndroidX {
-        const val activity = "1.2.0"
-        const val annotation = "1.1.0"
-        const val appCompat = "1.2.0"
-        const val core = "1.3.2"
-        const val lifecycle = "2.3.0"
+        const val activity = "1.3.0"
+        const val annotation = "1.2.0"
+        const val appCompat = "1.3.1"
+        const val core = "1.6.0"
+        const val lifecycle = "2.3.1"
 
         object View {
-            const val constraintLayout = "2.0.4"
+            const val constraintLayout = "2.1.0"
             const val coordinatorLayout = "1.1.0"
             const val drawerLayout = "1.1.1"
-            const val material = "1.3.0"
-            const val recyclerView = "1.1.0"
+            const val material = "1.4.0"
+            const val recyclerView = "1.2.1"
             const val swipeRefreshLayout = "1.1.0"
             const val viewPager2 = "1.0.0"
         }
 
         object Test {
-            const val junit = "1.1.2"
-            const val rules = "1.3.0"
-            const val espresso = "3.3.0"
+            const val junit = "1.1.3"
+            const val rules = "1.4.0"
+            const val espresso = "3.4.0"
         }
     }
 
     object Firebase {
-        const val classpath = "4.3.5"
-        const val analytics = "18.0.2"
+        const val classpath = "4.3.8"
+        const val analytics = "19.0.0"
 
         object Crashlytics {
-            const val classpath = "2.5.0"
-            const val crashlytics = "17.3.1"
+            const val classpath = "2.7.1"
+            const val crashlytics = "18.2.0"
         }
 
         object Perf {
-            const val classpath = "1.3.4"
-            const val perf = "19.1.1"
+            const val classpath = "1.4.0"
+            const val perf = "20.0.2"
         }
     }
 
-    const val hilt = "2.32-alpha"
+    const val hilt = "2.38.1"
 
     const val ask = "0.5.2"
 
     const val logger = "0.4.0"
 
-    const val mockito = "3.7.7"
+    const val mockito = "3.11.2"
 }
 
 object Dependencies {